WORK IN PROGRESS!
This branch introduces a script for validating modules. Aim is to
align all the modules to certain standards and to check against
that standard quickly.

Pushed this branch to show what's going on, but please DON'T MERGE.
Commits on this branch may change as the wisdom coming from writing
the code gives better ideas about what was coded earlier.